+++ tests/test_xml.py
@@ -523,6 +523,46 @@ class TestXMPPXMLGenerator(XMLTestCase):
         gen.endDocument()
         gen.flush()
 
+    def test_redeclaration_of_namespace_prefixes(self):
+        gen = xml.XMPPXMLGenerator(self.buf, short_empty_elements=True)
+        gen.startDocument()
+        gen.startPrefixMapping("ns0", "uri:foo")
+        gen.startElementNS(("uri:foo", "foo"), None, {})
+        gen.startPrefixMapping("ns0", "uri:bar")
+        gen.startElementNS(("uri:foo", "bar"), None, {})
+        gen.endElementNS(("uri:foo", "bar"), None)
+        gen.endPrefixMapping("ns0")
+        gen.endElementNS(("uri:foo", "foo"), None)
+        gen.endPrefixMapping("ns0")
+        gen.endDocument()
+        self.assertEqual(
+            b'<?xml version="1.0"?>'
+            b'<ns0:foo xmlns:ns0="uri:foo">'
+            b'<ns1:bar xmlns:ns0="uri:bar" xmlns:ns1="uri:foo"/>'
+            b'</ns0:foo>',
+            self.buf.getvalue()
+        )
+
+    def test_redeclaration_of_default_namespace(self):
+        gen = xml.XMPPXMLGenerator(self.buf, short_empty_elements=True)
+        gen.startDocument()
+        gen.startPrefixMapping(None, "uri:foo")
+        gen.startElementNS(("uri:foo", "foo"), None, {})
+        gen.startPrefixMapping(None, "uri:bar")
+        gen.startElementNS(("uri:foo", "bar"), None, {})
+        gen.endElementNS(("uri:foo", "bar"), None)
+        gen.endPrefixMapping(None)
+        gen.endElementNS(("uri:foo", "foo"), None)
+        gen.endPrefixMapping(None)
+        gen.endDocument()
+        self.assertEqual(
+            b'<?xml version="1.0"?>'
+            b'<foo xmlns="uri:foo">'
+            b'<ns0:bar xmlns="uri:bar" xmlns:ns0="uri:foo"/>'
+            b'</foo>',
+            self.buf.getvalue()
+        )
+
     def tearDown(self):
         del self.buf